+++ capi/examples/input-struct-init.c
@@ -1,9 +1,9 @@
-#include <string.h>
 #include "blazesym.h"
 
 int main(int argc, const char* argv[]) {
-  BLAZE_INPUT(blaze_inspect_elf_src, src,
+  blaze_inspect_elf_src src = {
+    .type_size = sizeof(src),
     .path = "/tmp/some/dir/test.bin",
     .debug_syms = true,
-  );
+  };
 }