+++ capi/src/lib.rs
@@ -14,8 +14,7 @@
 //! The library aims to provide forward compatibility with newer
 //! versions and backward compatibility with older ones. To make that
 //! happen, relevant types that are being passed to the library contain
-//! the `type_size` member that is to be set to the type's size. The
-//! `BLAZE_INPUT` macro can be used for convenient initialization:
+//! the `type_size` member that is to be set to the type's size, e.g.:
 //! ```c
 #![doc = include_str!("../examples/input-struct-init.c")]
 //! ```